chore: moving JS constants around for reuse (#9261)
