+++ changelog/chore-moving-constants-around
@@ -1,0 +1,5 @@
+Significance: patch
+Type: dev
+Comment: chore: moving JS constants around for reuse
+
+
